# PNG: Add SSE/AVX version of Sub, Up, Average and Paethfilters