… .github/ISSUE_TEMPLATE/commercial-bug-report.yml
… .github/ISSUE_TEMPLATE/oss-bug-report.yml
… tests/ImageSharp.Tests/Formats/Png/PngDecoderTests.cs
… tests/ImageSharp.Tests/Formats/Png/PngEncoderFilterTests.cs
… tests/ImageSharp.Tests/TestImages.cs
… tests/Images/Input/Png/AverageFilter4Bpp.png
… tests/Images/Input/Png/PaethFilter4Bpp.png
… tests/Images/Input/Png/SubFilter4Bpp.png