### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description

This PR adds a SSE2 version of the average filter (4 bytes per pixel only), which is used for decoding PNG's.
I tried a SSE2 version for 3 bytes per pixel also, but benchmarks did not show god results. Probably due to misaligned read/write.

edit:
PR now also includes SSE/AVX versions of Paeth, Sub and Up filter.

#### Benchmark results:

master:
```
BenchmarkDotNet=v0.13.0, OS=Windows 10.0.19043.1526 (21H1/May2021Update)
Intel Core i7-6700K CPU 4.00GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET SDK=6.0.200
  [Host]     : .NET 5.0.14 (5.0.1422.5710), X64 RyuJIT
  Job-OGQEMG : .NET 5.0.14 (5.0.1422.5710), X64 RyuJIT
|                      Method |     Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------------- |---------:|----------:|----------:|------:|------:|------:|----------:|
| 'Average-filtered PNG file' | 2.626 ms | 0.0179 ms | 0.0046 ms |     - |     - |     - |      3 KB |
```

PR:
```
BenchmarkDotNet=v0.13.0, OS=Windows 10.0.19043.1526 (21H1/May2021Update)
Intel Core i7-6700K CPU 4.00GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET SDK=6.0.200
  [Host]     : .NET 5.0.14 (5.0.1422.5710), X64 RyuJIT
  Job-OGQEMG : .NET 5.0.14 (5.0.1422.5710), X64 RyuJIT

Runtime=.NET 5.0  Arguments=/p:DebugType=portable  IterationCount=5
LaunchCount=1  WarmupCount=3

|                             Method |        Mean |     Error |   StdDev | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------------- |------------:|----------:|---------:|------:|--------:|-------:|------:|------:|----------:|
| 'Average-filtered PNG file (4bpp)' | 2,212.37 us | 14.979 us | 3.890 us | 55.63 |    5.27 |      - |     - |     - |      3 KB |
```

Testimage was:
![AverageFilter4Bpp](https://user-images.githubusercontent.com/38701097/155379201-7f29f128-c819-4ec5-8594-eda718806e73.png)