@brianpopow 2022-02-25
I have added SSE/AVX versions of the other filters, too.

Benchmark with random data of size 1024:

```
BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19043.1526 (21H1/May2021Update)
Intel Core i7-6700K CPU 4.00GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET SDK=6.0.200
  [Host]     : .NET 5.0.14 (5.0.1422.5710), X64 RyuJIT
  DefaultJob : .NET 5.0.14 (5.0.1422.5710), X64 RyuJIT


|      Method |        Mean |     Error |    StdDev |
|------------ |------------:|----------:|----------:|
|    UpScalar |   525.50 ns |  2.733 ns |  2.557 ns |
|      UpSse2 |    36.66 ns |  0.160 ns |  0.150 ns |
|      UpAvx2 |    19.11 ns |  0.074 ns |  0.069 ns |
|   SubScalar |   654.48 ns |  3.637 ns |  3.224 ns |
|     SubSse2 |   172.30 ns |  2.738 ns |  2.561 ns |
| PaethScalar | 6,643.09 ns | 23.076 ns | 21.585 ns |
|   PaethSse2 |   674.23 ns |  2.468 ns |  2.309 ns |
```